require and validate node version 6+
